feat: #80-API 매개변수 수정
